pyproject.toml
[tool.poetry]
name = "dofus-uniy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scapy = "^2.6.0"
protobuf = "^5.28.2"
pycryptodomex = "^3.20.0"
python-decouple = "^3.8"
pyautogui = "^0.9.54"
pillow = "^10.4.0"
clandestino = "^0.1.13"
clandestino-sqlite = "^0.1.3"
meeseeks-singleton = "^0.4.4"
opencv-python = "^4.10.0.84"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
